move function to seperate file
